% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.R
\name{clust}
\alias{clust}
\title{Cluster Analysis of Pollutant Profiles}
\usage{
clust(data, pollutants, method = "kmeans", k = 3)
}
\arguments{
\item{data}{Data frame containing pollutant profiles for stations. Columns must include pollutants.}

\item{pollutants}{Vector of pollutant names for analysis.}

\item{method}{Clustering method to use ("kmeans" or "hclust").}

\item{k}{Number of clusters (for k-means method).}
}
\value{
Prints cluster results and generates a dendrogram or cluster plot based on the chosen method.
}
\description{
This function performs cluster analysis on pollutant profiles using either k-means or hierarchical clustering.
}
\examples{
data <- load_data()
pollutants <- c("PM10", "PM2.5")
clust(data, pollutants, method = "kmeans", k = 4)

}
