% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correg.R
\name{correg}
\alias{correg}
\title{Correlation and Regression Analysis with Plotting}
\usage{
correg(data, pollutants)
}
\arguments{
\item{x}{Vector containing data for the first pollutant.}

\item{y}{Vector containing data for the second pollutant.}

\item{pollutant_x}{Name of the first pollutant.}

\item{pollutant_y}{Name of the second pollutant.}
}
\value{
Prints correlation and regression results and generates a scatter plot with a regression line.
}
\description{
This function performs correlation and regression analysis between two pollutants and generates a scatter plot with the regression line.
}
\examples{
data <- load_data()
pollutants <- c("SO2", "NO2", "O3", "CO", "PM10", "PM2.5")
correg(data, pollutants)

}
