% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2a_conv.R
\name{d2a_conv}
\alias{d2a_conv}
\title{Daily to Annual Pollution Conversion}
\usage{
d2a_conv(data, date, pollutants)
}
\arguments{
\item{data}{A data frame containing pollution data.}

\item{date}{Column name which contains date in year-month-day format.}

\item{pollutants}{Vector of pollutant names for analysis.}
}
\value{
A data frame with annual average pollutant levels.
}
\description{
The function can be used to convert daily pollution values to annual values.
}
\examples{
data <- load_data()
pollutants <- c("SO2", "NO2", "O3", "CO", "PM10", "PM2.5")
d2a_conv(data = data, date = "Date", pollutants = pollutants)

}
