% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_plots.R
\name{dm_plots}
\alias{dm_plots}
\title{Daily, Monthly Time Series Plot}
\usage{
dm_plots(data, date, pollutants)
}
\arguments{
\item{data}{A data frame containing pollution data.}

\item{date}{Column name containing date in year-month-day format.}

\item{pollutants}{Vector of pollutant names for analysis.}
}
\value{
Plots of daily, monthly, and annual time series for each pollutant.
}
\description{
This function can be used to plot time series plots for pollution data.
Daily, monthly, and annual values of the time series are plotted.
}
\examples{
data <- load_data()
dm_plots(data, "Date", pollutants = c("SO2", "NO2", "O3", "CO", "PM10", "PM2.5"))

}
