% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highpol.R
\name{highpol}
\alias{highpol}
\title{Location with the Highest Pollutant Concentrations}
\usage{
highpol(file, pollutants)
}
\arguments{
\item{file}{A data frame containing pollutant measurements.}

\item{pollutants}{A vector of pollutant names for which the highest concentrations are to be determined.}
}
\value{
A list containing pollutant names and their corresponding highest concentration locations.
}
\description{
This function takes pre-loaded data and finds the locations with the highest
concentrations of specific pollutants.
}
\examples{
file <- load_data
pollutants= c("CO", "SO2", "NO2", "O3")
pollu_locs <- highpol(file, pollutants)
for (pollutant in pollutants) {
  cat("Pollutant:", pollutant, "\n")
  cat("Address:", pollu_locs[[pollutant]]$address, "\n\n")
}

}
