% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Pollution Data}
\usage{
load_data()
}
\value{
A data frame containing pollution data.
}
\description{
This function loads pollution data from the package's data directory.
}
\examples{
pollution_data <- load_data()
print(head(pollution_data))
}
