% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matriplot.R
\name{matriplot}
\alias{matriplot}
\title{Create a Matrix Plot for Pollutant Data}
\usage{
matriplot(data, pollutants)
}
\arguments{
\item{data}{A data frame containing the pollutant data.}

\item{pollutants}{A vector of pollutant names to include in the matrix plot.}
}
\value{
A matrix plot.
}
\description{
This function creates a matrix plot to visualize the relationships between pollutants.
}
\examples{
data <- load_data()
pollutants <- c("SO2", "NO2", "O3", "CO", "PM10", "PM2.5")
matriplot(data, pollutants)

}
